+++ CHANGES.md
@@ -3,3 +3,5 @@
 ## v0.2.0
 
 - refactor(global): Improved the experience.
+- refactor(client): Use EventEmitter instead of simple hook point.
+- perf(encoder): Fixed the debuff of encoder.